Refactored tests
